test(runtime): extract in-process ACP launch helper (#273)
